comments implemented
