+++ integrations/zendesk/src/actions/index.ts
@@ -7,7 +7,7 @@ import { getTicketConversation } from './get-ticket-conversation'
 import { listAgents } from './list-agents'
 import { openTicket } from './open-ticket'
 import { setConversationRequester } from './set-conversation-requester'
-import { makeApiRequest } from './make-api-request'
+import { callApi } from './call-api'
 import * as bp from '.botpress'
 
 export default {
@@ -20,5 +20,5 @@ export default {
   getTicketConversation,
   setConversationRequester,
   openTicket,
-  makeApiRequest,
+  callApi,
 } satisfies bp.IntegrationProps['actions']
